Align numeric equipment names with backend
