+++ frontend/src/app/models/enums/numerique.enum.ts
@@ -1,10 +1,22 @@
 export enum NUMERIQUE_EQUIPEMENT {
-  OrdinateurPortable = 'ordinateurPortable',
-  OrdinateurFixe = 'ordinateurFixe',
-  Smartphone = 'smartphone',
-  Tablette = 'tablette',
-  Ecran = 'ecran',
-  Serveur = 'serveur',
-  Routeur = 'routeur',
-  Imprimante = 'imprimante'
+  ECRAN = 'ECRAN',
+  SWITCH_ROUTEUR_FIREWALL = 'SWITCH_ROUTEUR_FIREWALL',
+  ORDINATEUR_FIXE_HAUTE_PERFORMANCE = 'ORDINATEUR_FIXE_HAUTE_PERFORMANCE',
+  ORDINATEUR_FIXE = 'ORDINATEUR_FIXE',
+  TELEVISION_MOYENNE_30_49_POUCES = 'TELEVISION_MOYENNE_30_49_POUCES',
+  MODEM = 'MODEM',
+  SERVEUR = 'SERVEUR',
+  VIDEOPROJECTEURS = 'VIDEOPROJECTEURS',
+  AUTRE_NUMERIQUE_EN_EUROS = 'AUTRE_NUMERIQUE_EN_EUROS',
+  ORDINATEUR_PORTABLE = 'ORDINATEUR_PORTABLE',
+  ENCEINTES = 'ENCEINTES',
+  IMPRIMANTE_LASER = 'IMPRIMANTE_LASER',
+  TABLETTE = 'TABLETTE',
+  RACKS = 'RACKS',
+  TELEVISION_GRANDE_55_POUCES = 'TELEVISION_GRANDE_55_POUCES',
+  ECRAN_PUBLICITAIRE_2M2 = 'ECRAN_PUBLICITAIRE_2M2',
+  BAIE_DE_DISQUES = 'BAIE_DE_DISQUES',
+  PHOTOCOPIEURS = 'PHOTOCOPIEURS',
+  IMPRIMANTE_JET_ENCRE = 'IMPRIMANTE_JET_ENCRE',
+  SMARTPHONE = 'SMARTPHONE'
 }
