+++ frontend/src/app/models/numerique-equipment-labels.ts
@@ -1,12 +1,25 @@
 import { NUMERIQUE_EQUIPEMENT } from './enums/numerique.enum';
 
 export const numeriqueEquipmentLabels: Record<NUMERIQUE_EQUIPEMENT, string> = {
-  [NUMERIQUE_EQUIPEMENT.OrdinateurPortable]: 'Ordinateur portable',
-  [NUMERIQUE_EQUIPEMENT.OrdinateurFixe]: 'Ordinateur fixe',
-  [NUMERIQUE_EQUIPEMENT.Smartphone]: 'Smartphone',
-  [NUMERIQUE_EQUIPEMENT.Tablette]: 'Tablette',
-  [NUMERIQUE_EQUIPEMENT.Ecran]: 'Écran',
-  [NUMERIQUE_EQUIPEMENT.Serveur]: 'Serveur',
-  [NUMERIQUE_EQUIPEMENT.Routeur]: 'Routeur',
-  [NUMERIQUE_EQUIPEMENT.Imprimante]: 'Imprimante'
+  [NUMERIQUE_EQUIPEMENT.ECRAN]: 'Écran',
+  [NUMERIQUE_EQUIPEMENT.SWITCH_ROUTEUR_FIREWALL]: 'Switch/Routeur/Firewall',
+  [NUMERIQUE_EQUIPEMENT.ORDINATEUR_FIXE_HAUTE_PERFORMANCE]:
+    'Ordinateur fixe haute performance',
+  [NUMERIQUE_EQUIPEMENT.ORDINATEUR_FIXE]: 'Ordinateur fixe',
+  [NUMERIQUE_EQUIPEMENT.TELEVISION_MOYENNE_30_49_POUCES]: 'Télévision 30-49 pouces',
+  [NUMERIQUE_EQUIPEMENT.MODEM]: 'Modem',
+  [NUMERIQUE_EQUIPEMENT.SERVEUR]: 'Serveur',
+  [NUMERIQUE_EQUIPEMENT.VIDEOPROJECTEURS]: 'Vidéoprojecteurs',
+  [NUMERIQUE_EQUIPEMENT.AUTRE_NUMERIQUE_EN_EUROS]: 'Autre numérique (en euros)',
+  [NUMERIQUE_EQUIPEMENT.ORDINATEUR_PORTABLE]: 'Ordinateur portable',
+  [NUMERIQUE_EQUIPEMENT.ENCEINTES]: 'Enceintes',
+  [NUMERIQUE_EQUIPEMENT.IMPRIMANTE_LASER]: 'Imprimante laser',
+  [NUMERIQUE_EQUIPEMENT.TABLETTE]: 'Tablette',
+  [NUMERIQUE_EQUIPEMENT.RACKS]: 'Racks',
+  [NUMERIQUE_EQUIPEMENT.TELEVISION_GRANDE_55_POUCES]: 'Télévision ≥55 pouces',
+  [NUMERIQUE_EQUIPEMENT.ECRAN_PUBLICITAIRE_2M2]: 'Écran publicitaire 2m²',
+  [NUMERIQUE_EQUIPEMENT.BAIE_DE_DISQUES]: 'Baie de disques',
+  [NUMERIQUE_EQUIPEMENT.PHOTOCOPIEURS]: 'Photocopieurs',
+  [NUMERIQUE_EQUIPEMENT.IMPRIMANTE_JET_ENCRE]: "Imprimante jet d'encre",
+  [NUMERIQUE_EQUIPEMENT.SMARTPHONE]: 'Smartphone'
 };
