--- frontend/src/app/num.zip/numerique-equipment-labels.ts
@@ -1,9 +1,0 @@
-import { NUMERIQUE_EQUIPEMENT } from './enums/numerique.enum';
-
-export const NumeriqueEquipmentLabels: Record<NUMERIQUE_EQUIPEMENT, string> = {
-  [NUMERIQUE_EQUIPEMENT.ECRAN]: 'Écran',
-  [NUMERIQUE_EQUIPEMENT.ORDINATEUR]: 'Ordinateur',
-  [NUMERIQUE_EQUIPEMENT.SERVEUR]: 'Serveur',
-  [NUMERIQUE_EQUIPEMENT.ROUTEUR]: 'Routeur',
-  [NUMERIQUE_EQUIPEMENT.SWITCH]: 'Switch',
-};
