--- frontend/src/app/num.zip/numerique-saisie-donnees-page.component.html
@@ -1,100 +1,0 @@
-<div class="main-container">
-  <div class="left-panel">
-    <h2>Poste numérique</h2>
-
-    <p class="question-title">Disposez-vous de données sur le cloud ?</p>
-    <label><input type="radio" name="hasCloud" [(ngModel)]="donneesCloudDisponibles" [value]="true" (change)="updateData()" /> Oui</label>
-    <label><input type="radio" name="hasCloud" [(ngModel)]="donneesCloudDisponibles" [value]="false" (change)="updateData()" /> Non</label>
-
-    <div class="card" *ngIf="donneesCloudDisponibles">
-      <label for="cloudTraffic">Trafic cloud (Go)</label>
-      <input id="cloudTraffic" type="number" [(ngModel)]="traficCloud" (blur)="updateData()" />
-
-      <label for="tipUtilisateur">TIP utilisateur</label>
-      <input id="tipUtilisateur" type="number" [(ngModel)]="tipUtilisateur" (blur)="updateData()" />
-    </div>
-
-    <hr />
-
-    <details>
-      <summary><strong>Ajout d’un nouvel équipement numérique</strong></summary>
-
-      <div class="form-column">
-        <label>Type d'équipement</label>
-        <select [(ngModel)]="nouvelEquipement.equipement">
-          <option *ngFor="let entry of equipementOptions" [ngValue]="entry.value">
-            {{ entry.label }}
-          </option>
-        </select>
-
-        <label>Quantité</label>
-        <input type="number" [(ngModel)]="nouvelEquipement.nombre" />
-
-        <label>Durée d’amortissement (années)</label>
-        <input type="number" [(ngModel)]="nouvelEquipement.dureeAmortissement" />
-
-        <label>Émissions de GES connues ?</label>
-        <select [(ngModel)]="nouvelEquipement.emissionsGesPrecisesConnues">
-          <option [ngValue]="false" selected>Non</option>
-          <option [ngValue]="true">Oui</option>
-        </select>
-
-        <label *ngIf="nouvelEquipement.emissionsGesPrecisesConnues">Émissions de GES réelles (kgCO2/équipement)</label>
-        <input *ngIf="nouvelEquipement.emissionsGesPrecisesConnues" type="number" [(ngModel)]="nouvelEquipement.emissionsReellesParProduitKgCO2e" />
-      </div>
-
-      <button type="button" (click)="ajouterEquipement()">Ajouter cet équipement</button>
-    </details>
-
-    <div *ngIf="equipementsAjoutes.length > 0">
-      <h3>Équipements numériques ajoutés</h3>
-      <table class="data-table">
-        <thead>
-          <tr>
-            <th>Type</th>
-            <th>Quantité</th>
-            <th>Amortissement</th>
-            <th>GES Connus</th>
-            <th>GES Réels</th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr *ngFor="let equipement of equipementsAjoutes">
-            <td>{{ numeriqueEquipmentLabels[equipement.equipement as NUMERIQUE_EQUIPEMENT] }}</td>
-            <td>{{ equipement.nombre }}</td>
-            <td>{{ equipement.dureeAmortissement }}</td>
-            <td>{{ equipement.emissionsGesPrecisesConnues ? 'Oui' : 'Non' }}</td>
-            <td>{{ equipement.emissionsReellesParProduitKgCO2e ?? 'N/A' }}</td>
-          </tr>
-        </tbody>
-      </table>
-    </div>
-
-    <div *ngIf="equipementsAnciens.length > 0">
-      <h3>Équipements enregistrés les années précédentes</h3>
-      <table class="data-table">
-        <thead>
-          <tr>
-            <th>Type</th>
-            <th>Quantité</th>
-            <th>Amortissement</th>
-            <th>GES Connus</th>
-            <th>GES Réels</th>
-            <th>Année</th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr *ngFor="let equipement of equipementsAnciens">
-            <td>{{ numeriqueEquipmentLabels[equipement.equipement as NUMERIQUE_EQUIPEMENT] }}</td>
-            <td>{{ equipement.nombre }}</td>
-            <td>{{ equipement.dureeAmortissement }}</td>
-            <td>{{ equipement.emissionsGesPrecisesConnues ? 'Oui' : 'Non' }}</td>
-            <td>{{ equipement.emissionsReellesParProduitKgCO2e ?? 'N/A' }}</td>
-            <td>{{ equipement.anneeAjout }}</td>
-          </tr>
-        </tbody>
-      </table>
-    </div>
-  </div>
-</div>
-<app-save-footer [path]="'numeriqueOnglet'" (estTermineChange)="onEstTermineChange($event)"></app-save-footer>
