--- frontend/src/app/num.zip/numerique.enum.ts
@@ -1,7 +1,0 @@
-export enum NUMERIQUE_EQUIPEMENT {
-  ECRAN = 'ECRAN',
-  ORDINATEUR = 'ORDINATEUR',
-  SERVEUR = 'SERVEUR',
-  ROUTEUR = 'ROUTEUR',
-  SWITCH = 'SWITCH'
-}
